requirements.txt
pandas==2.3.3
Pillow==12.0.0
streamlit==1.51.0
streamlit==1.37.1
streamlit==1.22.0
